3.1-Рефакторинг-кода
