+++ src/view/film-details.js
@@ -1,10 +1,12 @@
+import {genresAdd} from "../viewing.js";
+
 const createGenresItem = (value) => {
   return (`<span class="film-details__genre">${value}</span>`);
 };
 
 export const createFilmPopup = (movie) => {
   const {name, rate, release, genre, duration, poster, descriptionFull, comments, director, writers, actors} = movie;
-  const genresAdd = genre.split(` `).map((value) => createGenresItem(value)).join(``);
+
   return `<section class="film-details">
   <form class="film-details__inner" action="" method="get">
     <div class="film-details__top-container">
@@ -53,7 +55,7 @@ export const createFilmPopup = (movie) => {
             </tr>
             <tr class="film-details__row">
               <td class="film-details__term">Genres</td>
-              <td class="film-details__cell">${genresAdd}</td>
+              <td class="film-details__cell">${genresAdd(genre, createGenresItem)}</td>
             </tr>
           </table>
           <p class="film-details__film-description">${descriptionFull}</p>
